deployment_requirements.txt
pandas
numpy
torch
torchvision
joblib
pillow
flask